backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
beanie==1.26.0
motor==3.3.2
pymongo==4.6.3
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
arq==0.26.1
redis==5.0.8
httpx==0.27.2
anthropic==0.34.2
python-multipart==0.0.9
email-validator==2.2.0
aiodocker==0.23.0
